Remove installation path selection
Drop the ability to select the installation path again, and just install it
into the default location (where the extension lives) on extension install.
The install screen still remains (with just a button) in case things go
wrong; that way we can still trigger a fallback install if the backend ends
up missing.

Signed-off-by: Mark Yen <mark.yen@suse.com>